refactor(program): replace old implementation
