+++ apps/bulldozer-programs/Anchor.toml
@@ -9,4 +9,4 @@ cluster = "localnet"
 wallet = "~/.config/solana/id.json"
 
 [scripts]
-test = "ts-mocha -p ./tsconfig.json -t 1000000 tests/instruction-account-v2.spec.ts --parallel"
+test = "ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts --parallel"
