--- apps/bulldozer-programs/programs/bulldozer/src/collections/instruction_program_account.rs
@@ -1,11 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[account]
-pub struct InstructionProgramAccount {
-    pub authority: Pubkey,
-    pub application: Pubkey,
-    pub instruction: Pubkey,
-    pub name: Vec<u8>,
-    // program associated to the account
-    pub program: Pubkey,
-}
