--- apps/bulldozer-programs/programs/bulldozer/src/collections/instruction_signer_account.rs
@@ -1,11 +1,0 @@
-use anchor_lang::prelude::*;
-use crate::enums::{MarkAttribute};
-
-#[account]
-pub struct InstructionSignerAccount {
-    pub authority: Pubkey,
-    pub application: Pubkey,
-    pub instruction: Pubkey,
-    pub name: Vec<u8>,
-    pub mark_attribute: MarkAttribute,
-}
