+++ apps/bulldozer-programs/programs/bulldozer/src/enums/mod.rs
@@ -1,11 +1,9 @@
-mod attribute_kind;
-mod attribute_modifier;
 mod account_kind;
 mod account_modifier;
-mod mark_attribute;
+mod attribute_kind;
+mod attribute_modifier;
 
-pub use attribute_kind::*;
-pub use attribute_modifier::*;
 pub use account_kind::*;
 pub use account_modifier::*;
-pub use mark_attribute::*;
+pub use attribute_kind::*;
+pub use attribute_modifier::*;
