--- apps/bulldozer-programs/programs/bulldozer/src/instructions/delete_instruction_basic_account.rs
@@ -1,14 +1,0 @@
-use anchor_lang::prelude::*;
-use crate::collections::{InstructionBasicAccount};
-
-#[derive(Accounts)]
-pub struct DeleteInstructionBasicAccount<'info> {
-    #[account(mut, close = authority, has_one = authority)]
-    pub account: Account<'info, InstructionBasicAccount>,
-    pub authority: Signer<'info>,
-}
-
-pub fn handler(_ctx: Context<DeleteInstructionBasicAccount>) -> ProgramResult {
-    msg!("Delete instruction basic account");
-    Ok(())
-}
