--- apps/bulldozer-programs/programs/bulldozer/src/instructions/delete_instruction_signer_account.rs
@@ -1,14 +1,0 @@
-use anchor_lang::prelude::*;
-use crate::collections::{InstructionSignerAccount};
-
-#[derive(Accounts)]
-pub struct DeleteInstructionSignerAccount<'info> {
-    #[account(mut, close = authority, has_one = authority)]
-    pub account: Account<'info, InstructionSignerAccount>,
-    pub authority: Signer<'info>,
-}
-
-pub fn handler(_ctx: Context<DeleteInstructionSignerAccount>) -> ProgramResult {
-    msg!("Delete instruction signer account");
-    Ok(())
-}
