+++ apps/bulldozer-programs/programs/bulldozer/src/instructions/mod.rs
@@ -4,27 +4,18 @@ pub mod create_collection_attribute;
 pub mod create_instruction;
 pub mod create_instruction_account;
 pub mod create_instruction_argument;
-pub mod create_instruction_basic_account;
-pub mod create_instruction_program_account;
-pub mod create_instruction_signer_account;
 pub mod delete_application;
 pub mod delete_collection;
 pub mod delete_collection_attribute;
 pub mod delete_instruction;
 pub mod delete_instruction_account;
 pub mod delete_instruction_argument;
-pub mod delete_instruction_basic_account;
-pub mod delete_instruction_program_account;
-pub mod delete_instruction_signer_account;
 pub mod update_application;
 pub mod update_collection;
 pub mod update_collection_attribute;
 pub mod update_instruction;
 pub mod update_instruction_account;
 pub mod update_instruction_argument;
-pub mod update_instruction_basic_account;
-pub mod update_instruction_program_account;
-pub mod update_instruction_signer_account;
 
 pub use create_application::*;
 pub use create_collection::*;
@@ -32,24 +23,15 @@ pub use create_collection_attribute::*;
 pub use create_instruction::*;
 pub use create_instruction_account::*;
 pub use create_instruction_argument::*;
-pub use create_instruction_basic_account::*;
-pub use create_instruction_program_account::*;
-pub use create_instruction_signer_account::*;
 pub use delete_application::*;
 pub use delete_collection::*;
 pub use delete_collection_attribute::*;
 pub use delete_instruction::*;
 pub use delete_instruction_account::*;
 pub use delete_instruction_argument::*;
-pub use delete_instruction_basic_account::*;
-pub use delete_instruction_program_account::*;
-pub use delete_instruction_signer_account::*;
 pub use update_application::*;
 pub use update_collection::*;
 pub use update_collection_attribute::*;
 pub use update_instruction::*;
 pub use update_instruction_account::*;
 pub use update_instruction_argument::*;
-pub use update_instruction_basic_account::*;
-pub use update_instruction_program_account::*;
-pub use update_instruction_signer_account::*;
