--- apps/bulldozer-programs/programs/bulldozer/src/instructions/update_instruction_basic_account.rs
@@ -1,21 +1,0 @@
-use anchor_lang::prelude::*;
-use crate::collections::{Collection, InstructionBasicAccount};
-use crate::enums::{MarkAttribute};
-use crate::utils::{vectorize_string};
-
-#[derive(Accounts)]
-#[instruction(name: String, mark_attribute: u8)]
-pub struct UpdateInstructionBasicAccount<'info> {
-    #[account(mut, has_one = authority)]
-    pub account: Box<Account<'info, InstructionBasicAccount>>,
-    pub collection: Box<Account<'info, Collection>>,
-    pub authority: Signer<'info>,
-}
-
-pub fn handler(ctx: Context<UpdateInstructionBasicAccount>, name: String, mark_attribute: u8) -> ProgramResult {
-    msg!("Update instruction basic account");
-    ctx.accounts.account.name = vectorize_string(name, 32);
-    ctx.accounts.account.collection = ctx.accounts.collection.key();
-    ctx.accounts.account.mark_attribute = MarkAttribute::from_index(mark_attribute)?;
-    Ok(())
-}
